pkg/bpf: Move unused tail call pruning logic to separate file
Move the logic for pruning unused tail calls from collection.go to
a new file.
